Graphing almost done
The graphing section is almost done, all that needs to be added is the
ability to copy and save the graphs, also possible
reorginization/deletion of the bottom left graph

noticed bugs: I put constraints and error messages on having a negative
date with the calendar and no date selected with the checkboxes, but it
seems to cause weird behaviors. If nothing is ever graphed, not
selecting either makes the upper left graph get larger for some reason.
If graphs were previously created it seems to add behaviors in even
though an invalid date was selected.