fixed something else
the firstime creation accidently filled the lists twice before graphing
them, I fixed it. Also I found out how to put as many labels on the x
axis for the totalbehaviors chart as I want, the code is commented out
in the beginning of the graph function. It should be easy to figure out
how many to put on the graph at once depending on how big it is, though
I would like Shauns opinion on how many he wants to see. If you graph
every day for the 30 days option(the smallest one) it gets a bit
crowded.